fix(chat): close the stale probed-token window after an endpoint write
A model-endpoint write swaps `config.models.llm` synchronously inside
`ModelEndpointService.load_all()`, but re-probes `/v1/models` in a
background task. Between the two the registry already describes the new
endpoint while `_max_model_tokens_by_name` still holds the previous
one's `max_model_len`, so a chat request landing in that window is
answered by the new endpoint and preflighted against the old endpoint's
limit — the preflight/answer divergence partition-scoped resolution
exists to prevent.

Invalidate the probed cache inline (`invalidate_max_model_tokens`) before
scheduling the refresh, so the window falls through to the endpoint's
admin-configured size or the global `llm_context` fallback — the
conservative "not probed yet" state — instead of another endpoint's
value. The slow part still runs off the request path, so a dead endpoint
can't stall an admin write.

Guard the publish with a generation counter: a refresh that was
invalidated while probing discards its results rather than republishing
pre-write values over the invalidation that just closed the window.

Also pass each endpoint's configured `timeout` to its probe instead of
`get_openai_models`' generic default. Probes run serially under the
lock, so an unreachable endpoint previously held every later refresh —
and therefore widened the empty-cache window — for longer than the admin
had asked for.

+++ openrag/api/dependencies/llm.py
@@ -4,7 +4,7 @@ from fastapi import HTTPException, status
 from openai import AsyncOpenAI
 
 
-async def get_openai_models(base_url: str, api_key: str, timeout: int = 30):
+async def get_openai_models(base_url: str, api_key: str, timeout: float = 30):
     async with AsyncOpenAI(base_url=base_url, api_key=api_key, timeout=timeout) as client:
         models_response = await client.models.list()
         return models_response.data
